taskReducer
