--- client/src/redux/TaskDataSlice.ts
@@ -1,31 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-import { fetchAllTasks } from "./actions/taskActions";
-
-const tasksSlice = createSlice({
-  name: "tasks",
-  initialState: {
-    items: [],
-    loading: false,
-    error: null as string | null,
-  },
-  reducers: {
-    // Any other synchronous reducers you might need
-  },
-  extraReducers: (builder) => {
-    builder
-      .addCase(fetchAllTasks.pending, (state) => {
-        state.loading = true;
-        state.error = null;
-      })
-      .addCase(fetchAllTasks.fulfilled, (state, action) => {
-        state.loading = false;
-        state.items = action.payload;
-      })
-      .addCase(fetchAllTasks.rejected, (state, action) => {
-        state.loading = false;
-        state.error = action.error.message || "Failed to fetch tasks";
-      });
-  },
-});
-
-export default tasksSlice;
