+++ client/src/redux/reducers/rootReducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from "@reduxjs/toolkit";
 import roomSlice from "./roomReducer";
-import tasksSlice from "../TaskDataSlice";
+import tasksSlice from "./taskReducer";
 
 const rootReducer = combineReducers({
   rooms: roomSlice.reducer,
