+++ client/src/redux/reducers/taskReducer.ts
@@ -1,0 +1,52 @@
+import { createSlice } from "@reduxjs/toolkit";
+import { fetchAllTasks } from "../actions/taskActions";
+import { Room } from "../../interfaces/Room";
+
+export interface TaskItem {
+  readonly _id: string;
+  name: string;
+  description?: string;
+  isCompleted?: boolean;
+  dueDate: number;
+  completedAt?: Date | null;
+  createdAt: Date;
+  updatedAt: Date;
+  room: Room;
+}
+
+export interface TaskState {
+  items: TaskItem[];
+  status: "idle" | "loading" | "succeeded" | "failed";
+  error: string | null;
+}
+
+export const initialState: TaskState = {
+  items: [],
+  status: "idle",
+  error: null as string | null,
+};
+
+const tasksSlice = createSlice({
+  name: "tasks",
+  initialState,
+  reducers: {
+    // Any other synchronous reducers you might need
+  },
+  extraReducers: (builder) => {
+    builder
+      .addCase(fetchAllTasks.pending, (state) => {
+        state.status = "loading";
+        state.error = null;
+      })
+      .addCase(fetchAllTasks.fulfilled, (state, action) => {
+        state.status = "succeeded";
+        state.items = action.payload;
+      })
+      .addCase(fetchAllTasks.rejected, (state, action) => {
+        state.status = "failed";
+        state.error = action.error.message || "Failed to fetch tasks";
+      });
+  },
+});
+
+export default tasksSlice;
